Initial commit
I've so far just added genconf and mkbash. My new mkbash is kinda my old
one on steroids, as it also creates scripts based on templates for at
least bash and python.

I even generated a new version of mkbash with mkbash, to ensure it
follows its own format.

On the other hand, genconf is intentionally distinct. I don't want it to
be user-friendly because I don't want anybody other than myself to use
it.